Add files via upload
Opinião: We stayed here last minute due to Storm Eunice forcing us to go to London a day early as it was by far the best value hotel in the Bloomsbury area. Like many we've stayed away from the hotel before in favour of fancier hotels closer to the centre of town as this hotel has a reputation for being very basic as one of the few central London hotels with rooms regularly within work hotel allowance limits.

We were surprised by how good this place actually is. Our room was 6035. Our room had a king sized bed (rare in central London) and loads of space - so much space it was obvious a sofa bed was supposed to be in here too but that didn't bother us. The bathroom had clearly been recently renovated and it couldn't have been done better.

The hotel corridors suggest a very run down place well past its best with printed paper direction signs but the moment you're in your room you realise the management here are trying to really bring it up to a higher standard. It's a solid 2-3 star hotel and a better standard than most chain budget hotels (think good bed but no toiletries). With the best will in the world this is not in any way a 4 star hotel but it could easily be so with little effort. My wife said the noise from the street was very loud through the night and the lack of a noisy air con unit to provide white noise to mask the sound from the street was frustrating but I slept through it all no problem.